+++ src/item/ItemSword.java
@@ -3,13 +3,13 @@ package item;
 import entity.Entity;
 
 public class ItemSword extends Item{
-	private int damage;
+	private double damage;
 	
-	public ItemSword(String name, int durability, int damage) {
+	public ItemSword(String name, int durability, double damage) {
 		super(name, durability);
 		this.damage = damage;
 	}
-	
+
 	public void attackEntity(Entity entity){
 		entity.setHealthPoints(entity.getHealthPoints() - damage);
 	}
